Splitting MSP into serial and MSP parts- stage1
